requirements.txt
torch
torchvision
opencv-python>=4.13.0
numpy
pillow
detectron2
Flask
Flask-Bootstrap
Flask-Uploads
torchmetrics
pyyaml
transformers
timm
tqdm
torchreid>=0.2.5
scipy
scikitlearn
tensorflow
ultralytics






